DD_TRACE_HEADER_TAGS update: Allow AspNetCore listener to set spans based on headers in the HTTP response
